fix(core): fix inconsistencies on `core`
