+++ packages/core/src/api/applicationCommands.ts
@@ -36,8 +36,8 @@ export class ApplicationCommandsAPI {
 	 *
 	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#get-global-application-commands}
 	 * @param applicationId - The application id to fetch commands for
-	 * @param query - The query options to use when fetching commands
-	 * @param options - The options to use when fetching commands
+	 * @param query - The query options for fetching commands
+	 * @param options - The options for fetching commands
 	 */
 	public async getGlobalCommands(
 		applicationId: Snowflake,
@@ -55,8 +55,8 @@ export class ApplicationCommandsAPI {
 	 *
 	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#create-global-application-command}
 	 * @param applicationId - The application id to create the command for
-	 * @param body - The data to use when creating the command
-	 * @param options - The options to use when creating the command
+	 * @param body - The data for creating the command
+	 * @param options - The options for creating the command
 	 */
 	public async createGlobalCommand(
 		applicationId: Snowflake,
@@ -75,7 +75,7 @@ export class ApplicationCommandsAPI {
 	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#get-global-application-command}
 	 * @param applicationId - The application id to fetch the command from
 	 * @param commandId - The command id to fetch
-	 * @param options - The options to use when fetching the command
+	 * @param options - The options for fetching the command
 	 */
 	public async getGlobalCommand(
 		applicationId: Snowflake,
@@ -93,7 +93,7 @@ export class ApplicationCommandsAPI {
 	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#edit-global-application-command}
 	 * @param applicationId - The application id of the command
 	 * @param commandId - The id of the command to edit
-	 * @param body - The data to use when editing the command
+	 * @param body - The data for editing the command
 	 * @param options - The options for editing the command
 	 */
 	public async editGlobalCommand(
@@ -129,7 +129,7 @@ export class ApplicationCommandsAPI {
 	 *
 	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#bulk-overwrite-global-application-commands}
 	 * @param applicationId - The application id to overwrite commands for
-	 * @param body - The data to use when overwriting commands
+	 * @param body - The data for overwriting commands
 	 * @param options - The options for overwriting commands
 	 */
 	public async bulkOverwriteGlobalCommands(
@@ -149,8 +149,8 @@ export class ApplicationCommandsAPI {
 	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#get-guild-application-commands}
 	 * @param applicationId - The application id to fetch commands for
 	 * @param guildId - The guild id to fetch commands for
-	 * @param query - The data to use when fetching commands
-	 * @param options - The options to use when fetching commands
+	 * @param query - The data for fetching commands
+	 * @param options - The options for fetching commands
 	 */
 	public async getGuildCommands(
 		applicationId: Snowflake,
@@ -170,8 +170,8 @@ export class ApplicationCommandsAPI {
 	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#create-guild-application-command}
 	 * @param applicationId - The application id to create the command for
 	 * @param guildId - The guild id to create the command for
-	 * @param body - The data to use when creating the command
-	 * @param options - The options to use when creating the command
+	 * @param body - The data for creating the command
+	 * @param options - The options for creating the command
 	 */
 	public async createGuildCommand(
 		applicationId: Snowflake,
@@ -192,7 +192,7 @@ export class ApplicationCommandsAPI {
 	 * @param applicationId - The application id to fetch the command from
 	 * @param guildId - The guild id to fetch the command from
 	 * @param commandId - The command id to fetch
-	 * @param options - The options to use when fetching the command
+	 * @param options - The options for fetching the command
 	 */
 	public async getGuildCommand(
 		applicationId: Snowflake,
@@ -212,8 +212,8 @@ export class ApplicationCommandsAPI {
 	 * @param applicationId - The application id of the command
 	 * @param guildId - The guild id of the command
 	 * @param commandId - The command id to edit
-	 * @param body - The data to use when editing the command
-	 * @param options - The options to use when editing the command
+	 * @param body - The data for editing the command
+	 * @param options - The options for editing the command
 	 */
 	public async editGuildCommand(
 		applicationId: Snowflake,
@@ -252,8 +252,8 @@ export class ApplicationCommandsAPI {
 	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#bulk-overwrite-guild-application-commands}
 	 * @param applicationId - The application id to overwrite commands for
 	 * @param guildId - The guild id to overwrite commands for
-	 * @param body - The data to use when overwriting commands
-	 * @param options - The options to use when overwriting the commands
+	 * @param body - The data for overwriting commands
+	 * @param options - The options for overwriting the commands
 	 */
 	public async bulkOverwriteGuildCommands(
 		applicationId: Snowflake,
@@ -313,8 +313,8 @@ export class ApplicationCommandsAPI {
 	 * @param applicationId - The application id to edit the permissions for
 	 * @param guildId - The guild id to edit the permissions for
 	 * @param commandId - The id of the command to edit the permissions for
-	 * @param body - The data to use when editing the permissions
-	 * @param options - The options to use when editing the permissions
+	 * @param body - The data for editing the permissions
+	 * @param options - The options for editing the permissions
 	 */
 	public async editGuildCommandPermissions(
 		userToken: string,
