+++ packages/core/src/api/invite.ts
@@ -11,8 +11,8 @@ export class InvitesAPI {
 	 *
 	 * @see {@link https://discord.com/developers/docs/resources/invite#get-invite}
 	 * @param code - The invite code
-	 * @param query - The options to use when fetching the invite
-	 * @param options - The options to use when fetching the invite
+	 * @param query - The options for fetching the invite
+	 * @param options - The options for fetching the invite
 	 */
 	public async get(code: string, query: RESTGetAPIInviteQuery = {}, { signal }: Pick<RequestData, 'signal'> = {}) {
 		return this.rest.get(Routes.invite(code), {
@@ -26,7 +26,7 @@ export class InvitesAPI {
 	 *
 	 * @see {@link https://discord.com/developers/docs/resources/invite#delete-invite}
 	 * @param code - The invite code
-	 * @param options - The options to use when deleting the invite
+	 * @param options - The options for deleting the invite
 	 */
 	public async delete(code: string, { reason, signal }: Pick<RequestData, 'reason' | 'signal'> = {}) {
 		await this.rest.delete(Routes.invite(code), { reason, signal });
