+++ packages/core/src/api/roleConnections.ts
@@ -17,7 +17,7 @@ export class RoleConnectionsAPI {
 	 *
 	 * @see {@link https://discord.com/developers/docs/resources/application-role-connection-metadata#get-application-role-connection-metadata-records}
 	 * @param applicationId - The id of the application to get role connection metadata records for
-	 * @param options - The options to use when fetching the role connection metadata records
+	 * @param options - The options for fetching the role connection metadata records
 	 */
 	public async getMetadataRecords(applicationId: Snowflake, { signal }: Pick<RequestData, 'signal'> = {}) {
 		return this.rest.get(Routes.applicationRoleConnectionMetadata(applicationId), {
@@ -31,7 +31,7 @@ export class RoleConnectionsAPI {
 	 * @see {@link https://discord.com/developers/docs/resources/application-role-connection-metadata#update-application-role-connection-metadata-records}
 	 * @param applicationId - The id of the application to update role connection metadata records for
 	 * @param body - The new role connection metadata records
-	 * @param options - The options to use when updating the role connection metadata records
+	 * @param options - The options for updating the role connection metadata records
 	 */
 	public async updateMetadataRecords(
 		applicationId: Snowflake,
