+++ packages/core/src/api/stageInstances.ts
@@ -18,7 +18,7 @@ export class StageInstancesAPI {
 	 * Creates a new stage instance
 	 *
 	 * @see {@link https://discord.com/developers/docs/resources/stage-instance#get-stage-instance}
-	 * @param body - The data to use when creating the new stage instance
+	 * @param body - The data for creating the new stage instance
 	 * @param options - The options for creating the new stage instance
 	 */
 	public async create(
